+++ tests/test_manager.py
@@ -1,5 +1,5 @@
 def test_manager_works():
-    from yucca.training.trainers.YuccaLightningManager import YuccaLightningManager
+    from yucca.yucca.training.trainers.YuccaManager import YuccaLightningManager
 
     manager = YuccaLightningManager()
     assert manager is not None
