+++ yucca/training/lightning_modules/__init__.py
@@ -1,0 +1,3 @@
+import os
+
+os.environ["OMP_NUM_THREADS"] = "1"
