--- yucca/training/trainers/alternative_trainers/YuccaManager_24GB.py
@@ -1,34 +1,0 @@
-from yucca.training.trainers.YuccaManager import YuccaManager
-
-
-class YuccaManager_24GB(YuccaManager):
-    def __init__(
-        self,
-        model,
-        model_dimensions: str,
-        task: str,
-        folds: str | int,
-        plan_id: str,
-        starting_lr: float = None,
-        loss_fn: str = None,
-        momentum: float = None,
-        continue_training: bool = False,
-        checkpoint: str = None,
-        finetune: bool = False,
-        fast_training: bool = False,
-    ):
-        super().__init__(
-            model,
-            model_dimensions,
-            task,
-            folds,
-            plan_id,
-            starting_lr,
-            loss_fn,
-            momentum,
-            continue_training,
-            checkpoint,
-            finetune,
-            fast_training,
-        )
-        self.max_vram = 24
