--- yucca/training/trainers/alternative_trainers/YuccaTrainer_ForegroundOnly.py
@@ -1,34 +1,0 @@
-from yucca.training.trainers.YuccaTrainer import YuccaTrainer
-
-
-class YuccaTrainer_ForegroundOnly(YuccaTrainer):
-    def __init__(
-        self,
-        model,
-        model_dimensions: str,
-        task: str,
-        folds: str | int,
-        plan_id: str,
-        starting_lr: float = None,
-        loss_fn: str = None,
-        momentum: float = None,
-        continue_training: bool = False,
-        checkpoint: str = None,
-        finetune: bool = False,
-        fast_training: bool = False,
-    ):
-        super().__init__(
-            model,
-            model_dimensions,
-            task,
-            folds,
-            plan_id,
-            starting_lr,
-            loss_fn,
-            momentum,
-            continue_training,
-            checkpoint,
-            finetune,
-            fast_training,
-        )
-        self.p_force_foreground = 1.0
